♻refactor #108: 조회용 서비스 IssueQueryController 추가
