+++ backend/src/test/java/codesquard/app/issue/service/IssueServiceTest.java
@@ -38,6 +38,8 @@ class IssueServiceTest extends IntegrationTestSupport {
 	@Autowired
 	private IssueService issueService;
 	@Autowired
+	private IssueQueryService issueQueryService;
+	@Autowired
 	private UserService userService;
 	@Autowired
 	private UserRepository userRepository;
@@ -67,7 +69,7 @@ class IssueServiceTest extends IntegrationTestSupport {
 		Long id = createIssue();
 
 		// when
-		IssueReadResponse issueReadResponse = issueService.get(id);
+		IssueReadResponse issueReadResponse = issueQueryService.get(id);
 
 		// then
 		assertThat(issueReadResponse.getTitle()).isEqualTo("Service");
@@ -103,7 +105,7 @@ class IssueServiceTest extends IntegrationTestSupport {
 		issueService.modifyStatus(issueModifyStatusRequest, id);
 
 		// then
-		assertThat(issueService.findById(id).getStatus().name()).isEqualTo(issueStatus);
+		assertThat(issueQueryService.findById(id).getStatus().name()).isEqualTo(issueStatus);
 	}
 
 	@DisplayName("이슈 상태를 수정시 이슈가 없다면 400 에러를 반환한다.")
@@ -147,7 +149,7 @@ class IssueServiceTest extends IntegrationTestSupport {
 		issueService.modifyTitle(issueModifyTitleRequest, id);
 
 		// then
-		assertThat(issueService.findById(id).getTitle()).isEqualTo(title);
+		assertThat(issueQueryService.findById(id).getTitle()).isEqualTo(title);
 	}
 
 	@DisplayName("이슈 내용을 수정한다.")
@@ -163,7 +165,7 @@ class IssueServiceTest extends IntegrationTestSupport {
 		issueService.modifyContent(issueModifyContentRequest, id);
 
 		// then
-		assertThat(issueService.findById(id).getContent()).isEqualTo(content);
+		assertThat(issueQueryService.findById(id).getContent()).isEqualTo(content);
 	}
 
 	@DisplayName("이슈 마일스톤을 수정한다.")
@@ -182,7 +184,7 @@ class IssueServiceTest extends IntegrationTestSupport {
 		issueService.modifyMilestone(issueModifyMilestoneRequest, id);
 
 		// then
-		assertThat(issueService.findById(id).getMilestoneId()).isEqualTo(milestoneId2);
+		assertThat(issueQueryService.findById(id).getMilestoneId()).isEqualTo(milestoneId2);
 	}
 
 	@DisplayName("이슈 책임자를 수정한다.")
@@ -197,7 +199,7 @@ class IssueServiceTest extends IntegrationTestSupport {
 		issueService.modifyAssignees(issueModifyAssigneesRequest, id);
 
 		// then
-		assertThat(issueService.findAssigneesById(id)).isEmpty();
+		assertThat(issueQueryService.findAssigneesById(id)).isEmpty();
 	}
 
 	@DisplayName("이슈 라벨을 수정한다.")
@@ -219,7 +221,7 @@ class IssueServiceTest extends IntegrationTestSupport {
 		issueService.modifyLabels(issueModifyLabelsRequest, id);
 
 		// then
-		List<IssueLabelResponse> label = issueService.findLabelsById(id);
+		List<IssueLabelResponse> label = issueQueryService.findLabelsById(id);
 		assertThat(label.get(0).getName()).isEqualTo(name);
 		assertThat(label.get(0).getBackground()).isEqualTo(background);
 		assertThat(label.get(0).getColor().getNameToLowerCase()).isEqualTo(color);
@@ -237,7 +239,7 @@ class IssueServiceTest extends IntegrationTestSupport {
 		issueService.modifyLabels(issueModifyLabelsRequest, id);
 
 		// then
-		List<IssueLabelResponse> label = issueService.findLabelsById(id);
+		List<IssueLabelResponse> label = issueQueryService.findLabelsById(id);
 		assertThat(label).isEmpty();
 	}
